added summary tab to org details, factored edit/remove domain buttons back into AdminDomains.js
